Squashed Bug in Row Click
Initially the click event was always registered by the entire row. As
such even clicking a button on the particular row had caused the page to
load a new routerlink and carry out the button's task as well. Hence
when a click is detected on a button on the row, the event is not
propogated to the row itself. Thus only carrying out the task of the
button.