Search bar widths increased (#60)
* Search bar widths increased

Search Bar Widths for each phase has been increased to fill the size of
the grid they are positioned in. Mainly linked each mat-form-field to a
css class and increased the widths through the css implementation.

Refer to  #54

* Moved the CSS Style Class to Global CSS file.

Moved the class that expands the search-bar's length to the global
styles.css to reduce repeated code.

* Removed Erroneous Code

Accidental Code that wasn't initially removed is now deleted.