Working profiles.json reading and data validation.
The app is now able to read from an internal profiles.json during the
dev phase as well as an external profiles.json within the same directory
as the app in production. This commit also includes some level of data
validation to ensure that the json file is setup correctly.