Testing Table Component.
Currently there is a copy of the Table Component in Phase3 and the
Issues-Pending Component in Phase2. These are currently present to tweak
the UI to make sure that everything is functioning and acting as it
should be.

A copy of the Table Component must also be placed in Phase1 to test it
further and once the tests are complete, replace all hard-coded tables
within the app-phases.