Table Restructure (#115)
* Refactor data-tables dir to issue-tables

* IssueTable Skeleton Implemented in Phase 2

* IssueTable Skeletal Component

* Working Basic Table Structure

* Lint Fix

* Working Prototype of Table Component

* Testing Table Component.

Currently there is a copy of the Table Component in Phase3 and the
Issues-Pending Component in Phase2. These are currently present to tweak
the UI to make sure that everything is functioning and acting as it
should be.

A copy of the Table Component must also be placed in Phase1 to test it
further and once the tests are complete, replace all hard-coded tables
within the app-phases.

* Phase1 Tests Added

* Phase 2 Table Copies Implemented

* Completed Table Component

Previously hard-coded tables have been replaced with the table
component. To begin testing functionality.

* Remove label.service.ts changes

* Assignee Style Fix

* Visual Fixes

* Further Visual Fixes

* Clickable Area Highlights

* Severity Label Fixes

* Action Click Issue

* Code Fixes

* Codestyle Fixes

* Jasmine Fix