Fixed issue where Duplicates were not displayed.
Duplicates now display correctly when assigned to an edited / new issue.
Also noticed an issue whereby the latest comment of an issue is not
fetched (needed to work in the case where a new issue is assigned a
duplicate and the duplicate is then removed.) So added the functionality
to retrieve the most recent comment that matches the required template.