BugReportingRepository will be automatically created for Students (#189)
* BugReport Repo will be automatically created.

The BugReporting Phase Repository will be automatically created and
setup if it is missing during the initial log-in. Since the users are
required to have this repo for the PE.

* Streamline Repo Creation Process

Repository Initialization works only for the Bug Reporting Phase and
only for Students in that phase. (So Tutors/Admins will not have a
random repository created in their account). Error messages have been
put in place to identify any potential faliures.

* Usage of ErrorCode Constant

* Lint Fix

* Artificial wait time for repo creation

* User Permission Requested prior to Repo Creation

The app now seeks the user's permission prior to initializing the
BugReportingPhase's repository in their account (should it be missing).

* Error Message Fix

* Method Description Fix