Refactor issue model raj (#190)
* Lint Fix: const instead of let

* Migrated TeamResponsePhase to use BaseIssue

createIssueModel has been modified to work with baseIssue for the
TeamResponsePhase.

* Empty commit

* Fix case sensitivity when checking user authorization (#184)

* Fix case sensitivity when checking user authorization

* Remove console.log

* BugReportingRepository will be automatically created for Students (#189)

* BugReport Repo will be automatically created.

The BugReporting Phase Repository will be automatically created and
setup if it is missing during the initial log-in. Since the users are
required to have this repo for the PE.

* Streamline Repo Creation Process

Repository Initialization works only for the Bug Reporting Phase and
only for Students in that phase. (So Tutors/Admins will not have a
random repository created in their account). Error messages have been
put in place to identify any potential faliures.

* Usage of ErrorCode Constant

* Lint Fix

* Artificial wait time for repo creation

* User Permission Requested prior to Repo Creation

The app now seeks the user's permission prior to initializing the
BugReportingPhase's repository in their account (should it be missing).

* Error Message Fix

* Method Description Fix

* Fix Team Response Parsing

* TeamResp. Mostly working, left comment error

* Refactor of Issue Model Creation for TeamResponse

IssueModel Creation for TeamResponse has been refactored to follow the
new mode. The description reading bug has also been fixed as of this
commit.

* Fixed Duplicate Info Display

Duplicate information was not being correctly displayed on toggle. Issue
has now been fixed.

* Fixed Severity Toggling

Severity Toggle Used to reset the issue in some cases. That issue has
been resolved.

* Fix Issue Duplicate Removing Description

* Fix same issue with Assignees