Comment editor enhancements (#263)
* TypeScript Fix

* Save Button disabled on Upload

Save Button will be disabled when an upload is initiated and re-enabled
when upload is complete. This is currently only working in the
Tester-Response phase.

* Submit Button Disabled During File Upload for all Phases

* Button Text Modification on Upload

When a file is uploaded, the button text is changed to reflect that the
button is disabled to allow for the uploading of a file.

* Change Uploading Button Text

* Fix App Break Mid Upload

Previously when a file (not image) is in mid-upload stage, changing to
preview and clicking the makdown-hyperlink causes the app to break / log
the user out in some cases. This has been fixed by simply letting the
mid upload text exclude the parenthesis at the end.