DuplicatedIssuesComponent: Update team response comment (#445)
Previously, when the duplicate status of an issue was removed,
only its "duplicate" label was removed. This created an erroneous
state where the issue would still have a
"Duplicate of (parent issue ID)" text in its team response comment,
when its duplicate label was removed. Its former parent issue
would also keep tracking it as a duplicate.

Let's fix this bug by triggering an update to the team response
comment as well, when the duplicate status of an issue is removed
(in addition to the update to the duplicate label).