Setup automated builds for CATcher releases (#534)
Our release process requires the production artifacts to be
manually uploaded to GitHub (or deployed to gh-pages,
in the case of the browser version). This can be automated.

Let's create a GitHub Actions workflow that
automatically builds and uploads the production artifacts
for all four mediums (browser, Linux, MacOS and Windows),
in a single click.