Checks if each Open Phase is valid + Test Updates
Original Checks for undefined repos in Session Data is removed,
this means that...

- Admins can define placeholder names for the various phases in
the application without clearly stating the associated repo name
upfront.

Checks that ensure each phase listed on the `openPhases` array is
valid and belongs to a pre-defined Phase has been added.

This caused some breaking with existing tests but they have since
been updated to reflect acceptable behaviours when working with
`assertSessionDataIntegrity()`.