Add E2E Testing to GHActions Linux Job (#588)
Let's update our GH Actions Workflow to run the E2E Tests
on the Linux Job. E2E Tests  will be carried out on Chrome and
Firefox.

NOTE:
The `webdriver-manager update` command must be run before
executing E2E Tests. Currently this ensures that drivers related
to Firefox are updated to the latest version to work with our
tests. In the future this allows us to test our E2E Tests with the
latest browser drivers.