Implement cached logging capability (#616)
Currently, users do not have a way to easily retrieve logs generated
by the web application. It is also not possible to retrieve logs from
prior sessions.

Let's cache the app's logs, and allow the user to download them in a
text file.

Log information will be cached on localStorage. It can be
downloaded via a button, on the application header.
The app will save logs for a maximum of 4 sessions.