Maintain session ID through phase change (#672)
When the app navigates to a different phase, the IssueService's
session id is set to "undefined". Subsequently, any new bug reports
will have a meaningless (undefined) session id in their hidden data.

The session id should remain unmodified, even when switching between
phases. This will provide more flexibility for using the app.

To achieve this,
let's add a boolean parameter to the IssueService's reset logic,
to check if a session id reset is necessary when resetting data within
the IssueService.