ProfileService: read profiles from client data repo (#661)
The app's default profiles are stored within the
codebase (in the AppConfig). Hence, updating the profiles
requires us to release / deploy a new version of CATcher.

Let's fetch the profiles from a client data repository instead.
This makes it possible to update the profiles
without modifying our codebase.

We can define a new ProfileService that uses a HTTP request
to fetch these profiles from the repo.