Markdown Links: Fix redirection href values  (#666)
Markdown links: Fix redirection href values

Our application crashes when an invalid link is rendered as HTML 
and the user clicks on the HTML Link. This is because there are no
checks done before the link is rendered. 

Let's open these links in a new tabs and propagate these errors there 
such that current users do not lose their current progress with respect 
to bug reports and other form data.