Change implementation of LabelService.getLabelDefinition()
LabelService.getLabelDefinition() only considers labelValue of
the Label being considered. This might result in poor scalability
in the case that more labels are added, of which they have the same
labelValue.

Let's improve the logic of LabelService.getLabelDefinition()
by considering both the labelCategory as well as labelValue
to ensure that the correct definition for the corresponding
Label is fetched and displayed.