Upgrade to Electron v11 (#717)
We are using Electron v9, which is no longer supported by the
Electron team. Let's incrementally adopt newer versions of
Electron, starting with v11.

To adopt Electron v11, all usages of the `remote` module have been
removed from ElectronService. This is because the Electron team
is planning to deprecate the `remote` module. This module is disabled
by default from Electron v10 onwards, marked for deprecation in v12,
and set for removal in v14.

The affected functionality has been re-implemented in the main
process code (main.ts and electron-utils/menu-bar.ts).

This change only involves functionality in the desktop app.
(specifically, the ability to open external links, clear cookies
and the "Inspect Element" feature).

Details on the deprecation of `remote` module can be found at:
- https://www.electronjs.org/docs/api/remote
- https://github.com/electron/electron/issues/21408
- https://www.electronjs.org/docs/breaking-changes#deprecated-remote-module