LabelService: Create only the needed labels (#738)
When synchronising the repository in the user's account for Bug Reporting
and Tester Response phases, CATcher automatically create labels that are
unnecessary and unused. (e.g. status.* and responses.* labels).
These labels are used in the team response phase.

Within LabelService, let's separate the labels required for different
phases so that only the relevant labels are created in the repository
 of the corresponding phase.