Show Label definition (#726)
Certain labels (e.g type.Feature_Flaw) are ambiguous to the user.

The user might not fully understand what the scope of the label is and needs 
the label reference from the CS2103T website to grasp its meaning. This 
becomes troublesome for the user as they are unable to easily access the label 
definitions. 

Let's make use of the Angular Dialogue Material to display more information 
about the label if the user chooses the read more about the label within the 
app itself.