Closed issue syncing (#734)
Remove closed issues during syncing of issues

Fetched issues during the GitHub API calls go through the 
createAndSaveIssueModal and updateLocalStore methods. 

Issues that do not appear in the subsequent API fetch call remain in the 
local store since no issue cleaning up of the local store is performed

Let's
- Check for outdated issues in the current local store
- Remove the outdated issues in the local store

Issues that have been closed or are outdated should be removed from the 
local store and no longer be displayed on CATcher

GitHub's Cache might return a 304 to indicate no update of resource 
required. In this scenario, no issues will be returned and it is handled inside 
the getOutdatedIssues method accordingly