Improve session selection (#778)
Improve Session Selection

Users would have to manually fill in sessions when logging into
CATcher which can be confusing for them if they are using CATcher
for the first time. 

Let's fill in the required session form fields automatically by
creating a Profile from the encoded URL given to them.