Fix cursor position after pasting image (#857)
Currently, the user may experience inconvenience after pasting an image
into the comment editor, since the cursor may jump to the image url
after a successful upload. This occurs if the user continues to type
after pasting the image.

Let's fix this issue by changing the logic of calculating the new
position of the cursor after uploading an image.