Merge Release V3.4.1 with Master  (#858)
* Include release branch in GH actions workflow (#806)

The GitHub Actions pipeline on CATcher runs only on PRs targeted to
the master branch. 

Let's modify the configuration settings for the coverage-report and 
github-actions workflow files such that they both run for PRs 
targeted for merge into the release branch.

* Fix Label Icon Display for Label Definitions (#804)

There are major css issues with regards to the info buttons for the 
labels in the CATcher app when creating or editing an issue. 

Let's fix the CSS layout of the entire label-definition select 
option by formatting the HTML elements into a grid.

* Update version

* Allow HTML Tags in Markdown text  (#761)

Currently, our markdown sanitizer blocks every HTML tag.

Let's modify the sanitization pipeline such that it only occurs 
during editing in the markdown comment editor.

* Update version to v3.4.1

* Fix Linter Errors

Co-authored-by: Anubhav <35621759+anubh-v@users.noreply.github.com>
Co-authored-by: Audrey Felicio Anwar <54733401+AudreyFelicio@users.noreply.github.com>