Add error handling for malformed issues / comments (#866)
Issues with no comments that match the Team Response or Tester Response
templates are treated silently, i.e. there is no warning or error
message displayed to the user. In fact, the issue does not appear in
CATcher. 

Let's check if issues have parse errors, and flag them as faulty issues.
Let's also create a new component in view-issue that shows that the
issue has parse errors, and direct the student to the teaching team. 
Lastly, let's create a Faulty Issues table in the Tester Response phase.