Create issue templates (#899)
As an Open Source Project, it becomes painful to see developers writing
bug reports which do not clarify what the bug is about. (Which was one
of the purposes of CATcher as well)

Let's come up with a few issue templates for Bug Reports / Feature
Requests such that issue reporters are guided to give good features /
clarify their issues before submitting.