Abstract code to find comments that match a template to the Template class (#904)
Abstract code to find comments that match a template to superclass

There is some duplicated code across 4 templates that deal with finding
a comment that conforms to the comment template. 

Let's abstract this code into the Template superclass and create a new
parseFailure attribute inside Template. Let's also use this opportunity
to standardise the variable names across the templates. Lastly, let's
use parseFailure in Issue.