Improve Duplicate Issue Search in Team Response Phase (#894)
Choosing an issue as the parent of a duplicate issue can be
challenging. Users can receive close to a hundred issues in
the Team Response Phase so they would need to scroll
through everything to find the correct issue.

Let's add a search bar in the selection dropdown menu for
users to search for the parent issue by name or ID. Also, let's
include the ID of the parent duplicate issue for issues that
has already been duplicated, to ease the finding of the parent
issue.