Use coroutine parsers for Issue Templates (#910)
Use coroutine parsers in Issue Templates

The way that issue comments are parsed to extract relevant information
e.g. duplicate issue number or the team chosen severity is complicated
and reliant on regex, which is not human friendly nor maintainable.

Let's use the coroutine parsers in arcsecond to parse the comments
instead. The coroutine parsers allow us to easily extract the
information that we need within the parser itself. For the more
complicated sections, let's implement parsers for those specific
sections as well.

Let's also write tests for these new parsers. 

Note that the eventual aim is to remove SectionalDependency, Section
and its related subclasses, hence parsers were also implemented for
the Tutor Moderation phase.