Integrate arcsecond parsers and remove Section (#920)
Integrate arcsecond parsers and remove Section

The arcsecond parsers are only used for checking if a comment or issue
matches the respective template, and extracting the relevant
information is still being done by Section and its related classes.

Let's use the parsers to extract the information instead, and remove
the dependencies on Section and its subclasses. This allows us to
completely remove regex from the parsing of Github issues and comments.
Let's also update the Issue class and the relevant tests.

Lastly, let's change the name of some folders to better reflect their contents.