Fix title overflow bug (#1013)
Fix bug where an extremely long title for an issue
overflows out of the screen if the title contains no spaces.

Currently, a long title that exceeds the screen width will
only be wrapped at points where there is a space. If there
are no spaces, then no line breaks can be introduced, which
causes the title to just overflow out of the screen width.

To fix this, lets add a CSS rule to allow line breaks to be
introduced anywhere in the title.