Remove electron dependencies  (#1153)
Refactor and remove uses of Electron from CATcher

There are many uses of Electron in CATcher, even with the CATcher Electron
app being deprecated.  As such, there are many instances of unused code,
which is bad practice. Removing instances of uses of Electron allows us to
trim the codebase and remove dead code.

Let's 
* Refactor and remove uses of Electron from CATcher
* Remove all uses of Electron in the `src` folder first.
* Remove uses of Electron from `package.json` and other files
* Delete folders such as `electron-utils` and files such as `postinstall` that
are used to initialise electron
* Remove and amend scripts in `package.json` that are involved in
building the electron app
* Edit `deployment-actions.yml` to remove pre-build actions from CD
workflow