Downgrade ubuntu version in playwright.yml (#1298)
Downgrade ubuntu version to v22.04 in playwright.yml

Playwright v1.33~ is incompatible with ubuntu-latest
(v24.x at the point of writing), causing the playwright
workflow to fail at installing browser stage.

Let's fix the ubuntu version to v22.04 (i.e. the latest
compatible ubuntu version).

Note that if we were to upgrade the playwright version
(instead of downgrading ubuntu version), it would
require us to upgrade node version to >= 18, which
is non-trivial.