Fix formats (#1303)
Fix incorrect format of commented templates.

Templates for the tester response phase's comments on GitHub found in 
the code comments are incorrect. The templates include an extra full 
stop after "Team Chose [ {{ tag }} ]" and "Originally [ {{ tag }} ]" and 
a new line after "**Reason for disagreement:**". 

This causes the issue to be filed under faulty issues when viewed in the 
Tester response phase on CATcher.

Let's fix the formats by removing the extra full stops and replacing the 
new line with a space.