Add new phase called bug trimming to CATcher (#1308)
This new implementation is needed to support the new phase of
CATcher.

Multiple phases have to be updated to support the phase. These files
are phase.model.ts, session.model.ts, permission.service.ts and
issue.service.ts.

The new implementation of the phase is similar to how previous
phases are implemented.

The new public_data setting for the new phase can be found in
PR #1308.