Add type to correctly process issues from the REST API (#1321)
Issue responses from the REST API are being accepted as the GithubIssue
class. This is incorrectly typed since the REST API returns a lower
cased state, while GithubIssue expects an upper-cased state.

This causes an inconsistency when updating the issues in
issue.service.ts with the response from these REST API queries, which
is done during close and reopen. This prevents us from filtering based
on the state correctly.

Let's add a new type called GithubRestIssue that accurately represents
the issue response from the REST API. We then add a factory method in
GithubIssue that creates a GithubIssue from a GithubRestIssue and
replace all incorrect GithubIssue usages with GithubRestIssue.